requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
plotly>=5.18.0
httpx>=0.26.0
cryptography>=41.0.0
PyJWT>=2.0.0
pyyaml>=6.0
requests>=2.31.0
snowflake-connector-python[pandas]>=3.6.0
tenacity>=8.2.0
python-dotenv>=1.0.0
pyarrow>=14.0.0,<18.0.0
numpy-financial>=1.0.0
openpyxl>=3.1.0
